Move Operator and SimpleOperator into one file.
Move classes to match the reference implementation. Rename files
accordingly. Specs are all green.
