+++ alpha/ruby/lib/plan_out.rb
@@ -1,6 +1,5 @@
-require_relative 'plan_out/random_operators'
+require_relative 'plan_out/op_random'
 require_relative 'plan_out/assignment'
-require_relative 'plan_out/experiment'
 require_relative 'plan_out/simple_experiment'
 require_relative 'plan_out/version'
