--- alpha/ruby/lib/plan_out/op_simple.rb
@@ -1,20 +1,0 @@
-require_relative 'operator'
-
-module PlanOut
-  class OpSimple < Operator
-    def execute(mapper)
-      @mapper = mapper
-      @parameters = {}
-
-      @args.each do |key, value|
-        @parameters[key] = mapper.evaluate(value)
-      end
-
-      simple_execute
-    end
-
-    def simple_execute
-      -1
-    end
-  end
-end
